requirements.txt
vosk
pvporcupine
pyttsx3
pytesseract
pyautogui
transformers
torch
pyaudio
numpy
pillow
opencv-python
mss
openai
pynput
cryptography
requests